kernel-3.x.x: netfilter: nf_nat: change sequence number adjustments to 32 bits, commit 2d89c68a upstream

+++ trunk/linux-3.0.x/include/net/netfilter/nf_nat.h
@@ -25,7 +25,7 @@ struct nf_nat_seq {
 	u_int32_t correction_pos;
 
 	/* sequence number offset before and after last modification */
-	int16_t offset_before, offset_after;
+	int32_t offset_before, offset_after;
 };
 
 /* Single range specification. */
